Move monitor code into monitor.go
